requirements.txt
molo.core
django-geojson [field]
django-leaflet==0.18.0
django-import-export
